main메서드 축소, UI클래스 병합
