+++ src/ProjectCode/UICustomActionListener.java
@@ -7,20 +7,18 @@ import java.awt.event.ActionListener;
 import java.awt.event.MouseEvent;
 import java.awt.event.MouseListener;
 
-public class CustomActionListener  extends Component implements ActionListener, MouseListener {
+public class UICustomActionListener extends Component implements ActionListener, MouseListener {
     private InsertNickname insertNickname;
-    private MessengerUI ui;
-    //private Rename rename;
+    private UIMessenger ui;
     private String nickname;
     private ClientProtocol op;
     public UIMain uiMain;
     ServerMain sm = new ServerMain();
 
-    public CustomActionListener(InsertNickname insertNickname, MessengerUI ui/*, Rename rename*/, UIMain uiMain, ClientProtocol op) {
+    public UICustomActionListener(InsertNickname insertNickname, UIMessenger ui, UIMain uiMain, ClientProtocol op) {
         this.insertNickname = insertNickname;
         this.ui = ui;
         this.uiMain = uiMain;
-        //this.rename = rename;
         this.op = op;
     }
 
@@ -59,30 +57,6 @@ public class CustomActionListener  extends Component implements ActionListener, 
                 op.sendMsg("Create#" + roomName); // 그룹창 생성
             }
         }
-        /*} else if (command.equals("rename_nickname")) {
-            rename.setVisible(true);
-        } else if (command.equals("rename")) {
-            JTextField renameField = rename.getRenameField();
-            String newNickname = renameField.getText();
-            if (newNickname.length() > 0 && newNickname.length() <= 10) {
-                String oldNickname = nickname;
-                nickname = newNickname;
-                ui.setTitle(nickname + "님의 대화창");
-                JTextArea roomDisplay = ui.getroomDisplay();
-                roomDisplay.setText(roomDisplay.getText().replace(oldNickname, nickname));
-                JTextArea msgDisplay = ui.getMsgDisplay();
-                msgDisplay.append(oldNickname + "님이 닉네임을 변경했습니다. 새로운 닉네임: " + nickname + "\n");
-                rename.setVisible(false);
-                ui.setVisible(true);
-            } else {
-                JOptionPane.showMessageDialog(rename, "닉네임은 1~10자로 입력해주세요.");
-            }
-        }*/
-    }
-
-    @Override
-    public void mouseClicked(MouseEvent e) {
-
     }
 
     @Override
@@ -99,19 +73,8 @@ public class CustomActionListener  extends Component implements ActionListener, 
             }
         }
     }
-
-    @Override
-    public void mouseReleased(MouseEvent e) {
-
-    }
-
-    @Override
-    public void mouseEntered(MouseEvent e) {
-
-    }
-
-    @Override
-    public void mouseExited(MouseEvent e) {
-
-    }
+    public void mouseClicked(MouseEvent e) {}
+    public void mouseReleased(MouseEvent e) {}
+    public void mouseEntered(MouseEvent e) {}
+    public void mouseExited(MouseEvent e) {}
 }
